Include verification of response body
